Add files via upload
 t-distributed stochastic neighbour embedding (tSNE) is a Machine Learning non-linear dimensionality reduction technique.
 
Perplexity- Defines balance between local and global aspects of your data. No. of points we want to preserve in local neighbourhood.

Sensitivity of tSNE to perplexity-
We should always run the algo for various value’s of perplexity.
If perplexity is large, clusters will tend to shrink into denser structures
 